work in simulation

+++ src/program_utils.c
@@ -5,7 +5,7 @@ long long get_timestamp()
     struct timeval tv;
 
     gettimeofday(&tv, NULL);
-    return (tv.tv_sec * 1000LL) + (tv.tv_usec / 1000LL);
+    return (tv.tv_sec * (unsigned long)1000) + (tv.tv_usec / 1000);
 }
 
 /*struct timeval (comes from sys/time.h) to hold the time
